Error codes support

+++ include/cry/base64.h
@@ -12,9 +12,9 @@
 extern "C" {
 #endif
 
-size_t cry_base64_encode(char *out, const char *in, size_t len);
+int cry_base64_encode(char *out, const char *in, size_t len);
 
-size_t cry_base64_decode(char *out, const char *in, size_t len);
+int cry_base64_decode(char *out, const char *in, size_t len);
 
 #ifdef __cplusplus
 }
